+++ src/bin/test_linux.rs
@@ -6,7 +6,6 @@ fn main() -> Result<()> {
     let rsys = Rsys::new();
     println!("KERNEL VERSION - {}", rsys::linux::kernel_release()?);
     println!("HOSTNAME - {}", rsys::linux::hostname()?);
-    println!("MEMORY - {:#?}", rsys.memory()?);
     println!("KERNEL_VERSION - {}", rsys.kernel_release()?);
     println!("{:#?}", rsys::linux::cpu::processor()?);
     Ok(())